Reapply coma model changes
