+++ docs/tudatpy/source/data.rst
@@ -20,6 +20,7 @@ to read inputs from other softwares and integrate them with Tudat(py).
    data/spacetrack
    data/processTrk234
    data/mission_data_downloader
+   data/coma_model
 
 
 Functions
