+++ docs/tudatpy/source/dynamics/environment_setup/atmosphere.rst
@@ -56,8 +56,10 @@ Functions
 
 .. autosummary::
 
+   empty_wind_model
    constant_wind_model
    custom_wind_model
+   coma_wind_model
    exponential_predefined
    exponential
    nrlmsise00
@@ -69,11 +71,17 @@ Functions
    custom_four_dimensional_constant_temperature
    scaled_by_constant
    scaled_by_function
+   coma_model_from_poly_data
+   coma_model_from_stokes_data
 
+
+.. autofunction:: tudatpy.dynamics.environment_setup.atmosphere.empty_wind_model
 
 .. autofunction:: tudatpy.dynamics.environment_setup.atmosphere.constant_wind_model
 
 .. autofunction:: tudatpy.dynamics.environment_setup.atmosphere.custom_wind_model
+
+.. autofunction:: tudatpy.dynamics.environment_setup.atmosphere.coma_wind_model
 
 .. autofunction:: tudatpy.dynamics.environment_setup.atmosphere.exponential_predefined
 
@@ -96,6 +104,10 @@ Functions
 
 .. autofunction:: tudatpy.dynamics.environment_setup.atmosphere.scaled_by_function
 
+.. autofunction:: tudatpy.dynamics.environment_setup.atmosphere.coma_model_from_poly_data
+
+.. autofunction:: tudatpy.dynamics.environment_setup.atmosphere.coma_model_from_stokes_data
+
 
 Classes
 -------
@@ -104,9 +116,11 @@ Classes
 .. autosummary::
 
    WindModelSettings
+   EmptyWindModelSettings
    ConstantWindModelSettings
    CustomWindModelSettings
    AtmosphereSettings
+   ComaSettings
    ExponentialAtmosphereSettings
    CustomConstantTemperatureAtmosphereSettings
    ScaledAtmosphereSettings
@@ -119,6 +133,10 @@ Wind Model Settings
 
 .. autoclass:: tudatpy.dynamics.environment_setup.atmosphere.WindModelSettings
    :members:
+
+.. autoclass:: tudatpy.dynamics.environment_setup.atmosphere.EmptyWindModelSettings
+   :members:
+   :show-inheritance:
 
 .. autoclass:: tudatpy.dynamics.environment_setup.atmosphere.ConstantWindModelSettings
    :members:
@@ -134,6 +152,10 @@ Atmosphere Settings
 
 .. autoclass:: tudatpy.dynamics.environment_setup.atmosphere.AtmosphereSettings
    :members:
+
+.. autoclass:: tudatpy.dynamics.environment_setup.atmosphere.ComaSettings
+   :members:
+   :show-inheritance:
 
 .. autoclass:: tudatpy.dynamics.environment_setup.atmosphere.ExponentialAtmosphereSettings
    :members:
