+++ src/tudat/astro/reference_frames/aerodynamicAngleCalculator.cpp
@@ -105,8 +105,18 @@ void AerodynamicAngleCalculator::update( const double currentTime, const bool up
     // Get current body-fixed state.
     if( !( currentTime == currentTime_ ) )
     {
-        currentBodyFixedGroundSpeedBasedState_ = bodyFixedStateFunction_( );
-        currentRotationFromCorotatingToInertialFrame_ = rotationFromCorotatingToInertialFrame_( );
+        currentRelativeState_ = relativeStateFunction_( );
+        currentRotationToLocalFrame_ = rotationToLocalFrameFunction_( );
+        currentRotationMatrixToLocalFrame_ = currentRotationToLocalFrame_.toRotationMatrix( );
+        currentRotationMatrixDerivativeToLocalFrame_ =
+                rotationMatrixDerivativeToLocalFrameFunction_ ? rotationMatrixDerivativeToLocalFrameFunction_( ) : Eigen::Matrix3d::Zero( );
+
+        currentRotationFromCorotatingToInertialFrame_ = currentRotationToLocalFrame_.inverse( );
+        currentBodyFixedGroundSpeedBasedState_.segment( 0, 3 ) =
+                currentRotationMatrixToLocalFrame_ * currentRelativeState_.segment( 0, 3 );
+        currentBodyFixedGroundSpeedBasedState_.segment( 3, 3 ) =
+                currentRotationMatrixToLocalFrame_ * currentRelativeState_.segment( 3, 3 ) +
+                currentRotationMatrixDerivativeToLocalFrame_ * currentRelativeState_.segment( 0, 3 );
 
         Eigen::Vector3d sphericalCoordinates =
                 coordinate_conversions::convertCartesianToSpherical< double >( currentBodyFixedGroundSpeedBasedState_.segment( 0, 3 ) );
@@ -115,8 +125,7 @@ void AerodynamicAngleCalculator::update( const double currentTime, const bool up
         currentAerodynamicAngles_[ latitude_angle ] = mathematical_constants::PI / 2.0 - sphericalCoordinates( 1 );
         currentAerodynamicAngles_[ longitude_angle ] = sphericalCoordinates( 2 );
 
-        // Compute wind velocity vector (used in both cases)
-        Eigen::Vector3d localWindVelocity = Eigen::Vector3d::Zero( );
+        currentLocalWindVelocity_.setZero( );
         if( windModel_ != nullptr )
         {
             // Determine whether to use radius or altitude based on wind model configuration
@@ -132,15 +141,13 @@ void AerodynamicAngleCalculator::update( const double currentTime, const bool up
                 altitudeOrRadius = shapeModel_->getAltitude( currentBodyFixedGroundSpeedBasedState_.segment( 0, 3 ) );
             }
 
-            localWindVelocity = getRotationQuaternionBetweenFrames( windModel_->getAssociatedFrame( ), corotating_frame ) *
+            currentLocalWindVelocity_ = getRotationQuaternionBetweenFrames( windModel_->getAssociatedFrame( ), corotating_frame ) *
                     windModel_->getCurrentBodyFixedCartesianWindVelocity(
                             altitudeOrRadius,
                             currentAerodynamicAngles_[ longitude_angle ],
                             currentAerodynamicAngles_[ latitude_angle ],
                             currentTime );
         }
-
-        currentLocalWindVelocity_ = localWindVelocity;
 
         // Compute airspeed-based velocity vector
         if( includeAtmosphericRotation_ )
@@ -149,47 +156,15 @@ void AerodynamicAngleCalculator::update( const double currentTime, const bool up
             // The body-fixed velocity already includes the rotational component
             // Airspeed = groundspeed - wind
             currentBodyFixedAirspeedBasedState_ = currentBodyFixedGroundSpeedBasedState_;
-            currentBodyFixedAirspeedBasedState_.segment( 3, 3 ) -= localWindVelocity;
+            currentBodyFixedAirspeedBasedState_.segment( 3, 3 ) -= currentLocalWindVelocity_;
         }
         else
         {
-            // NOTE: IDEs may falsely flag this block as "unreachable code" because all C++ wind model
-            // constructors default includeCorotation to true. However, this block IS reachable when users
-            // set include_corotation=False via Python bindings:
-            //
-            // Flow diagram when co-rotation is disabled:
-            //   Python: include_corotation=False (in custom_wind_model() or coma_settings.wind_settings)
-            //       ↓
-            //   C++ WindModel: includeCorotation_ = false
-            //       ↓
-            //   createFlightConditions.cpp: reads flag from wind model
-            //       ↓
-            //   AerodynamicAngleCalculator: includeAtmosphericRotation_ = false
-            //       ↓
-            //   THIS BLOCK EXECUTES: computes airspeed = groundspeed - Ṙ*r - wind
-            //
-            // Atmospheric rotation disabled: atmosphere does NOT co-rotate with the body
-            // The body-fixed groundspeed velocity contains: v_bodyfixed = Ṙ*r + R*v_inertial
-            // For non-rotating atmosphere, we need airspeed = R*v_inertial (velocity relative to inertial atmosphere)
-            // Therefore: v_airspeed = v_bodyfixed - Ṙ*r - wind
-
-            // Compute rotational velocity component Ṙ*r_inertial
-            // Note: transformStateToFrameFromRotations uses Ṙ*r_inertial, not Ṙ*r_bodyfixed
-            Eigen::Vector3d rotationalVelocity = Eigen::Vector3d::Zero( );
-            if( rotationMatrixDerivativeToLocalFrameFunction_ )
-            {
-                // First convert position from body-fixed to inertial frame
-                Eigen::Vector3d r_inertial = currentRotationFromCorotatingToInertialFrame_ *
-                                             currentBodyFixedGroundSpeedBasedState_.segment( 0, 3 );
-                // Then compute Ṙ*r_inertial (rotational velocity in body-fixed frame)
-                Eigen::Matrix3d rotationMatrixDerivative = rotationMatrixDerivativeToLocalFrameFunction_( );
-                rotationalVelocity = rotationMatrixDerivative * r_inertial;
-            }
-
-            // Compute airspeed by removing rotational velocity and wind
+            const Eigen::Vector3d rotationalVelocity =
+                    currentRotationMatrixDerivativeToLocalFrame_ * currentRelativeState_.segment( 0, 3 );
             currentBodyFixedAirspeedBasedState_ = currentBodyFixedGroundSpeedBasedState_;
             currentBodyFixedAirspeedBasedState_.segment( 3, 3 ) -= rotationalVelocity;
-            currentBodyFixedAirspeedBasedState_.segment( 3, 3 ) -= localWindVelocity;
+            currentBodyFixedAirspeedBasedState_.segment( 3, 3 ) -= currentLocalWindVelocity_;
         }
 
         // Calculate vertical <-> aerodynamic <-> body-fixed angles if neede.
